+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [4.1.0] - 2019-10-07
 ### Changed
  - Using new GenericBaseUtils to standardize crestron device setup and teardown
  - Using CresnetSettingsUtils to standardize cresnet settings serialization